// projects/adf-candidates/core/dialog/interfaces/dialog.interface.ts
/*!
 * @license
 * Copyright 2019 Alfresco, Inc. and/or its affiliates.
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *       http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */

export interface DialogData {
    title?: string;
    subtitle?: string;
    messages?: string[];
}

export interface HumanReadableChoice<T> {
    choice: T;
    title?: string;
    subtitle?: string;
    color?: 'primary' | 'accent' | 'default';
    spinnable?: boolean;
}
export interface MultipleChoiceDialogData<T> extends DialogData {
    choices: HumanReadableChoice<T>[];
}